+++ HW3 Design/wildlife_tracker/animal_management/animal.py
@@ -1,21 +1,19 @@
-from typing import Any, Optional
+from typing import Any, Optional, List
 
 class Animal:
 
-    #pass
     def __init__(self,
                 animal_id: int,
+                species: str,
                 health_status: Optional[str] = None,
                 age: Optional[int] = None) -> None:
         self.animal_id = animal_id
         self.health_status = health_status
         self.age = age
+        self.species = species
 
-    # def get_animal_by_id(animal_id: int) -> Optional[Animal]:
-    #     pass
-
-    def get_animal_details(animal_id) -> dict[str, Any]:
+    def get_animal_details(self) -> dict[str, Any]:
         pass
 
-    def update_animal_details(animal_id: int, **kwargs: Any) -> None:
+    def update_animal_details(self, **kwargs: Any) -> None:
         pass
