+++ HW3 Design/wildlife_tracker/formal_specification.py
@@ -2,7 +2,7 @@ from typing import Any, List, Optional
 from wildlife_tracker.animal_management.animal import Animal
 from wildlife_tracker.habitat_management.habitat import Habitat
 from wildlife_tracker.migration_tracking.migration import Migration
-from wildlife_tracker.migration_tracking.migration import MigrationPath
+from wildlife_tracker.migration_tracking.migration_path import MigrationPath
 
 age: Optional[int] = None
 animal_id: int
