+++ HW3 Design/wildlife_tracker/habitat_management/habitat.py
@@ -1,5 +1,7 @@
 from typing import Any, List, Optional
 
+from wildlife_tracker.animal_management.animal import Animal
+
 class Habitat:
 
     def __init__(self,
